Change format of CHANGELOG to Markdown
This moves `CHANGELOG` to `CHANGELOG.md` and makes its format Markdown.

Plain text has a certain aesthetic appeal, but I've found it to be
self-defeating in the vast majority of cases. Most people are reading
the log from GitHub rather than through a local editor, and having real
links to pull requests and richer formatting available are very nice
features.