Fix various Rustfmt problems in the build
The unstable features I was using are still unstable (they appear to be
permanently so), and the situation's gotten worse in that Rustfmt seems
to not really be available in nightly builds anymore, which was the only
way that unstable features could be used.

For now, stop using those unstable features and stop running Rustfmt in
nightly, preferring only the stable channel instead.