Upgrade version of dependencies + fix `time` breakages (#76)
To keep things somewhat current, upgrade all dependencies to latest
versions.

The hardest by far is `time`, which has broken its entire API surface
area, _again_. We make some fairly substantial code changes to
accommodate that.

I also changed the store API to use `u64`s instead of `i64`s since
they're working with nanoseconds. I realize in retrospect that this
change might not be needed since Redis works with a signed integer type,
but it improves the interface slightly in that a missing value is
returned as a `None` instead of `-1`, which is more Rust-like.