Include `Cargo.lock`
Include `Cargo.lock` as requested in #75.

I think I originally omitted it so that no one was stuck to specific
dependency versions that really didn't make any difference whatsoever,
but I guess that's not that big of a problem, and the lock file has some
nominal benefit for reproducibility/security.

Fixes #75.